Update to Java 8
 Change dependencies to build against latest red5 to upgrade to java 8

+++ bbb-screenshare/app/deploy.sh
@@ -9,7 +9,9 @@
 sbt clean
 sbt compile
 sbt package
-sudo rm -r /usr/share/red5/webapps/screenshare
+if [[ -d /usr/share/red5/webapps/screenshare ]]; then
+    sudo rm -r /usr/share/red5/webapps/screenshare
+fi
 sudo cp -r target/webapp/ /usr/share/red5/webapps/screenshare
 
 
@@ -19,11 +21,11 @@ sudo cp ~/dev/bigbluebutton/bbb-screenshare/app/target/webapp/WEB-INF/lib/bbb-sc
  ~/dev/bigbluebutton/bbb-screenshare/app/target/webapp/WEB-INF/lib/akka-* \
  ~/dev/bigbluebutton/bbb-screenshare/app/target/webapp/WEB-INF/lib/config-1.3.0.jar \
  ~/dev/bigbluebutton/bbb-screenshare/app/target/webapp/WEB-INF/lib/gson-1.7.1.jar \
- ~/dev/bigbluebutton/bbb-screenshare/app/target/webapp/WEB-INF/lib/spring-webmvc-4.1.8.RELEASE.jar  \
+ ~/dev/bigbluebutton/bbb-screenshare/app/target/webapp/WEB-INF/lib/spring-webmvc-4.2.5.RELEASE.jar  \
   /usr/share/red5/webapps/screenshare/WEB-INF/lib/
 
 
-sudo mkdir /usr/share/red5/webapps/screenshare/WEB-INF/classes
+#sudo mkdir /usr/share/red5/webapps/screenshare/WEB-INF/classes
 cd /usr/share/red5/webapps/screenshare/WEB-INF/classes/
 sudo jar -xf ../lib/bbb-screenshare-akka_2.11-0.0.1.jar
 sudo rm /usr/share/red5/webapps/screenshare/WEB-INF/lib/bbb-screenshare-akka_2.11-0.0.1.jar
